backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pillow
numpy
